+++ application.nix
@@ -178,5 +178,7 @@ rec {
       # Set a low default timeout when stopping services, to prevent the Windows 95 shutdown experience
       systemd.extraConfig = "DefaultTimeoutStopSec=15s";
 
+      playos.hardening.enable = true;
+
     };
 }